+++ src/components/content/index.ts
@@ -4,3 +4,5 @@ export { LearnSidebar } from './LearnSidebar';
 export { RelatedArticles } from './RelatedArticles';
 export { MDXContent } from './MDXContent';
 export { mdxComponents } from './mdx';
+export { LearnSearch } from './LearnSearch';
+export { BlogSearch } from './BlogSearch';